% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{add_complementary}
\alias{add_complementary}
\title{Add several entries, when only one is not NA.}
\usage{
add_complementary(z)
}
\arguments{
\item{z}{a vector containing at most only one non-NA value.}
}
\value{
row vector of the same size as entry matrix.
}
\description{
\code{add_complementary} return the only non NA value of a given vector.
}
\details{
This function is used in function \code{merge_complementary_vectors}
}
\keyword{internal}
