% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{add_pheno_error_variance}
\alias{add_pheno_error_variance}
\title{Add the pheno errors to a variance matrix}
\usage{
add_pheno_error_variance(varr, pheno_error)
}
\arguments{
\item{varr}{matrix of variance, result of function \code{compute_variance_covariance}}

\item{pheno_error}{a SymmetricMatrix of dimension p}
}
\value{
matrix of variance covariance plus errors
}
\description{
\code{add_pheno_error_variance} takes the  (n+m)*p variance covariance,
 matrix of vec(X), and add the pheno (independent) errors.
}
\keyword{internal}
