% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{add_zero_length_tips}
\alias{add_zero_length_tips}
\title{Add artificial tips}
\usage{
add_zero_length_tips(phylo)
}
\arguments{
\item{phylo:}{a phylogenetic tree of class phylo}
}
\value{
A phylogenetic tree of class phylo
}
\description{
\code{add_zero_length_tips} creates a new tree, where extra tips are
added to each tip, with a zero-length branch.
}
\keyword{internal}
