% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{allocate_shifts_from_regimes}
\alias{allocate_shifts_from_regimes}
\title{Allocation of shifts to edges}
\usage{
allocate_shifts_from_regimes(phylo, regimes)
}
\arguments{
\item{phylo}{: imput tree}

\item{regimes}{: vector of size (ntaxa + nNodes) of the regimes of each node
and tip.}
}
\value{
vector of edges numbers where the shifts are

10/10/14 - Initial release
}
\description{
\code{allocate_shifts_from_regimes} returns the position of the shifts induced
by the allocation of the regimes. Only works in an "infinite site" model.
}
\details{
This function uses function fun on each row of matrix of edges.
}

