% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{allocate_subset_node.simulate}
\alias{allocate_subset_node.simulate}
\title{Iteration allocation}
\usage{
allocate_subset_node.simulate(node, array, value)
}
\arguments{
\item{node}{node on which to slice}

\item{array}{structure to be sliced}

\item{value}{value to be attributed to the slice}
}
\value{
array: array p x Nnode x 2 (BM), with slice corresponding to node filled with value
}
\description{
\code{allocate_subset_node.simulate} slice the data correctly and allocate the right part.
To be used in function UpdateDown.
}
\keyword{internal}
