% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{check.selection.strength}
\alias{check.selection.strength}
\title{Check selection strength}
\usage{
check.selection.strength(process, selection.strength = NA,
  eps = 10^(-6), ...)
}
\arguments{
\item{process}{: "BM" or "OU"}

\item{selection.strength}{the selection strength parameter (if OU)}

\item{eps}{the tolerance for the selection strength}
}
\value{
character : "BM" or "OU"
}
\description{
\code{check.selection.strength} checks, if the process is an OU, if the
selection strength is not too low, in which case the process is replaced
with a BM.
}
\details{
This function return a process in a form of a character. If the entry
process is "BM", then nothing is done. If it is "OU", then the verification
of the selection strength is done.
}
\keyword{internal}
