% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{check_parsimony_ism}
\alias{check_parsimony_ism}
\title{Check wether an allocation of the shifts is parsimonious, 
in the "infinite site model".}
\usage{
check_parsimony_ism(tree, edges, ...)
}
\arguments{
\item{tree}{phylogenetic tree}

\item{edges}{a vector of edges of the tree, where the shifts are}

\item{...}{possibly, a list giving the descendant tips of each edge}
}
\value{
boolean : TRUE if the allocation is parsimonious.
}
\description{
\code{check_parsimony} take a vector of shifts edges, and check wether the number of 
groups of the tips induced by this allocation is exactly the number of shifts plus one.
}
\details{
This function computes explicitely the clustering of the tips, using 
function \code{check_parsimony}.
By default, this function uses \code{enumerate_tips_under_edges} to compute 
the list of tips under each edge, but a list can be provided (if many tests are done).
}

