% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{clusters_from_shifts}
\alias{clusters_from_shifts}
\title{Clustering associated to a shift allocation, assuming no homoplasy.}
\usage{
clusters_from_shifts(tree, edges,
  part.list = enumerate_tips_under_edges(tree))
}
\arguments{
\item{tree}{phylogenetic tree}

\item{edges}{a vector of edges of the tree, where the shifts are}

\item{part.list}{a list giving the descendant tips of each edge}
}
\value{
list of size n+m-1, entry i is the vector of tips bellow edge i.
}
\description{
\code{clusters_from_shifts} take a vector of shifts edges, and gives the
clustering of the tips induced by them, in a "no homoplasy" model (i.e. no
convergence is allowed).
}
\details{
By default, this function uses \code{\link{enumerate_tips_under_edges}} to compute 
the list of tips under each edge.
}

