% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{compute_actualization_matrix_ultrametric}
\alias{compute_actualization_matrix_ultrametric}
\title{Compute Matrix W of actualization (Ultrametric case)}
\usage{
compute_actualization_matrix_ultrametric(tree, selection.strength,
  times_shared = compute_times_ca(tree))
}
\arguments{
\item{tree}{a phylogenetic tree.}

\item{selection.strength}{the selection strength of the process.}

\item{times_shared}{a matrix, result of function \code{compute_times_ca}.}
}
\value{
Matrix of size p*nedges
}
\description{
\code{compute_actualization_matrix_ultrametric} computes a squares  p*nedges bloc diagonal
matrix of the (I_p - exp(-A * (h - t_pa(j))))_{j node}.
}
\details{
Carreful: the root is not taken into account in this function.
}
\keyword{internal}

