% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{compute_betas}
\alias{compute_betas}
\title{Computation of the optimal values at nodes and tips.}
\usage{
compute_betas(phylo, optimal.value, shifts)
}
\arguments{
\item{phylo}{: imput tree}

\item{optimal.value}{the optimal value at the root of the tree}

\item{shifts}{position and values of the shifts}
}
\value{
Vector of size (ntaxa + nNodes) of the ptimal values at the tips
of the tree.
}
\description{
\code{compute_betas} computes the optimal values at the nodes and tips of the
tree, given the value at the root and the list of shifts occuring in the tree.
}
\details{
This function uses function \code{recursionDown} for recursion on the tree, 
with \code{init.compute_betas} for the initialization of the vector, and 
\code{update.compute_betas} for its actualization.
}

