% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{compute_betas_from_shifts}
\alias{compute_betas_from_shifts}
\title{Computation of the optimal values at nodes and tips.}
\usage{
compute_betas_from_shifts(phylo, optimal.value, shifts)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}

\item{optimal.value}{the optimal value at the root of the tree.}

\item{shifts}{position and values of the shifts .}
}
\value{
Vector of size (ntaxa + nNodes) of the ptimal values at the tips
of the tree.
}
\description{
\code{compute_betas_from_shifts} computes the optimal values at the nodes and tips of the
tree, given the value at the root and the list of shifts occuring in the tree.
It assumes an OU model.
}
