% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{compute_diff_exp.BM}
\alias{compute_diff_exp.BM}
\title{Compute differences of expectations between node and parent.}
\usage{
compute_diff_exp.BM(phylo, conditional_law_X)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{conditional_law_X}{result of function \code{compute_E}}
}
\value{
matrix p x nEdges containing, for each edge e finishing at node i,
the quantity E[Z_i|Y]-E[Z_pa(i)|Y].
}
\description{
\code{compute_diff_exp} compute the differences of conditional expectations between all the
nodes and their parents.
}
\keyword{internal}

