% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{compute_dist_phy}
\alias{compute_dist_phy}
\title{Phylogenetic Distances}
\usage{
compute_dist_phy(phy)
}
\arguments{
\item{phy}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}
}
\value{
a matrix of phylogenetic distances, ordered as the tips of the
tree. The matrix is of type \code{\link[Matrix]{symmetricMatrix-class}}.
}
\description{
\code{compute_dist_phy} computes the phylogenetic distances d_ij between all the
tips i, j.
}
\details{
Tis function relies on \code{ape} function
\code{\link[ape]{dist.nodes}}.
}

