% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_fixed_moments}
\alias{compute_fixed_moments}
\title{Compute fixed moments for E step.}
\usage{
compute_fixed_moments(times_shared, ntaxa)
}
\arguments{
\item{times_shared}{matrix of the tree, result of function
\code{compute_times_ca}}
}
\value{
F_means the matrix to use for actualization of means.

F_vars the matrix to use for the actualization of variances.
}
\description{
\code{compute_fixed_moments} compute the fixed matrices used in E step when
there is no missing data.
}
\keyword{internal}

