% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{compute_K_max}
\alias{compute_K_max}
\title{Maximal number of shifts allowed}
\usage{
compute_K_max(ntaxa, kappa = 0.9)
}
\arguments{
\item{ntaxa}{the number of tips}

\item{kappa}{a real strictly bellow 1.}
}
\value{
K_max the maximal number of shifts allowed.
}
\description{
\code{compute_K_max} computes the quantity 
min(floor(kappa * ntaxa / (2 + log(2) + log(ntaxa))), ntaxa - 7))
that is the maximal dimention allowed to get theoretical garenties during
the selection model, when using the procedure defined by Baraud et al (2009)
}
\details{
See Baraud et al (2009)
}
\keyword{internal}

