% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_residuals.simple}
\alias{compute_residuals.simple}
\title{Residuals}
\usage{
compute_residuals.simple(phylo, Y_data_vec, sim, Sigma_YY_chol_inv, miss)
}
\arguments{
\item{phylo}{Input tree.}

\item{sim}{(list) : result of function \code{simulate}.}

\item{Y_data}{: vector indicating the data at the tips.}

\item{Sigma_YY_inv}{: invert of the variance-covariance matrix of the data.}
}
\description{
\code{compute_residuals.simple} computes the residuals after the fit of the
data in the simple case where the inverse of the variance matrix is given.
}
\details{
This function takes parameters sim and Sigma_YY_inv from  
\code{compute_mean_variance.simple}. It uses function \code{extract_simulate_internal}
to extract the needed quantities from these objects.
}
\keyword{internal}
