% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{compute_shifts_from_betas}
\alias{compute_shifts_from_betas}
\title{Computation of shifts from the vector of optimal values}
\usage{
compute_shifts_from_betas(phylo, betas)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}

\item{betas}{vector of size (ntaxa + nNodes) of the optimal values at each
node and tip.}
}
\value{
vector of shifts.
}
\description{
\code{compute_shifts_from_betas} computes the list of shifts corresponding
to the vector of optimal values on nodes.
}
\details{
This function uses function fun on each row of matrix of edges.
}

