% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{compute_stationary_variance}
\alias{compute_stationary_variance}
\title{Compute the stationary variance matrix}
\usage{
compute_stationary_variance(variance, selection.strength)
}
\arguments{
\item{variance}{the variance (rate matrix) of the process.}

\item{selection.strength}{the selection strength (alpha) matrix of the 
process.}
}
\value{
A positive definite Matrix of class \code{\link[Matrix]{dpoMatrix-class}}.
}
\description{
\code{compute_stationary_variance} computes the stationary variance matrix of
an OU process.
}

