% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{compute_sum_var_diff}
\alias{compute_sum_var_diff}
\title{Compute weighted sum of var_diff}
\usage{
compute_sum_var_diff(phylo, var_diff)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{var_diff}{result of function \code{compute_var_diff.BM}}
}
\value{
matrix p x p
}
\description{
\code{compute_sum_var_diff} computes sum_\{e edge\} ell_j * Var[X_j - X_pa(j) | Y]
}
\keyword{internal}
