% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{compute_times_ca}
\alias{compute_times_ca}
\title{Common Ancestors Times}
\usage{
compute_times_ca(phy)
}
\arguments{
\item{phy}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}
}
\value{
a matrix of times of shared evolution, ordered as the tips of the
tree. The matrix is of type \code{\link[Matrix]{symmetricMatrix-class}}.
}
\description{
\code{compute_times_ca} computes the times t_ij between the root and the common
ancestor of two tips i, j.
}
\details{
This function relies on \code{ape} functions
\code{\link[ape]{node.depth.edgelength}} and \code{\link[ape]{mrca}}.
}
