% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_tree_correlations_matrix.BM}
\alias{compute_tree_correlations_matrix.BM}
\title{Matrix of tree-induced correlations for the BM}
\usage{
compute_tree_correlations_matrix.BM(times_shared, params_old, ...)
}
\arguments{
\item{times_shared}{times of shared ancestry of all nodes and tips, result of function
\code{compute_times_ca}}
}
\value{
times_shared
}
\description{
\code{compute_tree_correlations_matrix.BM} returns times_shared its provided argument.
}
\keyword{internal}

