% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{compute_var_diff.BM}
\alias{compute_var_diff.BM}
\title{Compute variances of differences between nodes and parents.}
\usage{
compute_var_diff.BM(phylo, conditional_law_X)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{conditional_law_X}{result of function \code{compute_E}}
}
\value{
array p x p x Nedge containing, for each edge e finishing at node i,
the quantity Var[Z_i-Z_pa(i)|Y].
}
\description{
\code{compute_var_diff} computes variances of differences between all the
nodes and their parents.
}
\keyword{internal}
