% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{compute_var_M.BM}
\alias{compute_var_M.BM}
\title{Computation of the variance.}
\usage{
compute_var_M.BM(phylo, var_diff, diff_exp, edges_max, random.root, mu)
}
\arguments{
\item{phylo}{Tree}

\item{var_diff}{variances of differences result of function \code{compute_var_diff.BM}}

\item{diff_exp}{differences of expectations result of function \code{compute_diff_exp.BM}}

\item{edges_max}{Edges where the shifts occur result of function \code{segmentation.BM}}
}
\value{
a p x p matrix : the computed variance
}
\description{
\code{compute_var_M.BM} finds the variance that is the maximum of likelihood
}
\details{
Given the variances, the costs0 and the edges where the shifts occurs, the
computation of the maximum of likelihood in the variance is simple.
}
\keyword{internal}

