% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_variance_block_diagonal.BM}
\alias{compute_variance_block_diagonal.BM}
\title{Tips Variances for the BM}
\usage{
compute_variance_block_diagonal.BM(times_shared, params_old, ntaxa, ...)
}
\arguments{
\item{times_shared}{times of shared ancestry of all nodes and tips, result 
of function \code{compute_times_ca}}

\item{params_old}{(list) : old parameters to be used in the E step}
}
\value{
p * p * ntaxa array with ntaxa variance matrices
}
\description{
\code{compute_variance_block_diagonal.BM} computes the n p*p variance
matrices of each tip vector.
}
\keyword{internal}
