% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_variance_covariance.OU.nonsym}
\alias{compute_variance_covariance.OU.nonsym}
\title{Complete variance covariance matrix for OU}
\usage{
compute_variance_covariance.OU.nonsym(times_shared, distances_phylo,
  params_old, ...)
}
\arguments{
\item{times_shared}{times of shared ancestry of all nodes and tips, result 
of function \code{compute_times_ca}}

\item{params_old}{(list) : old parameters to be used in the E step}
}
\value{
matrix of variance covariance for the OU
}
\description{
\code{compute_variance_covariance.OU} computes the (n+m)*p squared variance
covariance matrix of vec(X).
}
\keyword{internal}
