% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{conditional_expectation_log_likelihood.OU.OLD}
\alias{conditional_expectation_log_likelihood.OU.OLD}
\title{Expectation conditional to the tips of the completed log-likelihood.
This is an old version.}
\usage{
conditional_expectation_log_likelihood.OU.OLD(phylo, conditional_law_X, sigma2,
  mu, alpha)
}
\arguments{
\item{phylo}{Input tree.}

\item{conditional_law_X}{result of function \code{compute_E.OU}, containing
first and second moments.}

\item{sigma2}{variance of the OU.}

\item{mu}{mean of the root.}

\item{alpha}{selection strength of the OU.}
}
\value{
double : value of the function
}
\description{
\code{conditional_expectation_log_likelihood.OU} computes the expectation
 conditional to the tips of the completed log-likelihood, given the 
 first and second order moments of the nodes, and the parameters of the 
 OU process.
}
\details{
This function uses functions \code{compute_var_diff.OU} 
and \code{compute_diff_exp.OU} in the process. Carefull : only works if the
root is stationary, and shifts at nodes.
}
\keyword{internal}

