% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{correspondenceEdges}
\alias{correspondenceEdges}
\title{Correspondence between edges numbers}
\usage{
correspondenceEdges(edges, from, to)
}
\arguments{
\item{edges}{vector of index of edges in the tree "from"}

\item{from}{initial input tree}

\item{to}{aimed output tree}
}
\value{
vector of index of edges in the tree "to"

26/05/14
}
\description{
\code{correspondenceEdges} takes edges numbers on an input tree, and gives back their
corresponding numbers on the output tree.
}

