% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{enlight}
\alias{enlight}
\alias{enlight.PhyloEM}
\title{Make the result lighter}
\usage{
enlight(x)

\method{enlight}{PhyloEM}(x)
}
\arguments{
\item{x}{an object of class \code{\link{PhyloEM}}.}
}
\value{
Same as entry, lighter.
}
\description{
\code{enlight.PhyloEM} takes an object of class \code{\link{PhyloEM}},
and returns the same object, without saving the quantities that can be easily
re-computed using function \code{\link{imputed_traits.PhyloEM}}.
}
\details{
The resulting object can be much lighter, saving a lot of memory space, but each
call to the function \code{\link{imputed_traits.PhyloEM}} will be longer. As 
function \code{\link{plot.PhyloEM}} relies on this function, this makes the 
plotting also longer.
This has the same effect as setting the option "\code{light_result=TRUE}" in the
call of \code{\link{PhyloEM}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{PhyloEM}: \code{\link{PhyloEM}} object
}}

\seealso{
\code{\link{PhyloEM}}, \code{\link{imputed_traits.PhyloEM}},
\code{\link{plot.PhyloEM}}
}
