% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{enumerate_parsimony}
\alias{enumerate_parsimony}
\title{Enumerate all the possible regime allocations, given a culstering 
of the tips.}
\usage{
enumerate_parsimony(phylo, clusters = rep(1, length(phylo$tip.label)))
}
\arguments{
\item{phylo}{Input tree.}

\item{clusters}{a vector representing the group of each tip.}
}
\value{
an S3 object of class "\code{enumerate_parsimony}", with;
\describe{
\item{nbrReconstructions}{an object of class "\code{parsimonyCost}", result
of function \code{\link{parsimonyCost}}.}
\item{allocations}{a list of size nNode + ntaxa. Each entry i of the list
represents the solutions for the subtree starting at node i. It is a list with
nclus entries, each entry being a matrix. A line of the kth matrix for the
ith node is one possible allocation of the shifts, starting with regime k
for node i.}
}
}
\description{
\code{enumerate_parsimony} enumerate all the equivalent allocation of the 
regimes in the tree, a clustering of the tips being given. The number of such
equivalent regimes is given by \code{parsimonyNumber} (which is faster).
}
\details{
This functin does a recursion up the tree, using functions 
\code{init.enumerate_parsimony} for the initialization at the tips, 
\code{updateUp_list} for the effective recursion on the tree,
and \code{update.enumerate_parsimony} for the actualisation of the parameters.
The function \code{extract_enumerate_parsimony} furnishes the result in a nice
form of a matrix (for any subtree).
}

