% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{enumerate_tips_under_edges}
\alias{enumerate_tips_under_edges}
\title{Tips descendants of nodes.}
\usage{
enumerate_tips_under_edges(tree)
}
\arguments{
\item{tree}{phylogenetic tree, class \code{\link[ape]{phylo}}.}
}
\value{
list of size Nedge, entry i is the vector of tips bellow edge i.
}
\description{
\code{enumerate_tips_under_edges} gives, for each edge of the tree, the labels
of the tips that have this edge as an ancestor.
}
\details{
This function uses function \code{\link[ape]{prop.part}} from package \code{ape}.
}
