% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{equivalent_shifts}
\alias{equivalent_shifts}
\title{Find all equivalent shifts allocations and values.}
\usage{
equivalent_shifts(phylo, params, T_tree = incidence.matrix(phylo),
  part.list = enumerate_tips_under_edges(phylo), times_shared = NULL)
}
\arguments{
\item{phylo}{a phylogenetic tree, of class \code{\link[ape]{phylo}}.}

\item{params}{an object of class \code{params_process}, result inference by
function \code{\link{PhyloEM}}, or constructed throught function
\code{\link{params_process}}}

\item{T_tree}{(optional) matrix of incidence of the tree, result of function 
\code{\link{incidence.matrix}}}

\item{part.list}{(optional) list of partition of the tree, result of function
\code{\link{enumerate_tips_under_edges}}.}

\item{times_shared}{(optional) a matrix, result of function
\code{\link{compute_times_ca}}.}

\item{...}{further arguments to be passed to \code{\link[ape]{plot.phylo}}.}
}
\value{
object of class \code{equivalent_shifts}, whith entries:
\describe{
\item{eq_shifts_edges}{matrix of equivalent shifts, result of function
\code{\link{equivalent_shifts_edges}}}
\item{shifts_and_betas}{matrix of corresponding shifts values}
\item{phylo}{the entry phylogenetic tree}
\item{p}{the dimention}
}
}
\description{
\code{equivalent_shifts} computes the equivalent shifts positions using
\code{\link{equivalent_shifts_edges}}, and then there corresponding values,
assuming an ultrametric tree.
}
\details{
This function is only valid for ultrametric trees, and for models: BM, OU with
fixed root or stationary root.
}
\seealso{
\code{\link{plot.equivalent_shifts}},
\code{\link{equivalent_shifts_edges}}
}

