% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{equivalent_shifts_values}
\alias{equivalent_shifts_values}
\title{Find values given edges. OU stationary case. Ultrametric tree.}
\usage{
equivalent_shifts_values(phylo, shifts, beta_0,
  eq_shifts_edges = equivalent_shifts_edges(phylo, shifts$edges), T_tree_ac)
}
\arguments{
\item{phylo}{a phylogenetic tree.}

\item{shifts}{a list of positions and values of original shifts.}

\item{beta_0}{value of the original optimal value at the root.}

\item{eq_shifts_edges}{matrix, result of function \code{equivalent_shifts_edges}.}

\item{T_tree_ac}{matrix of incidence of the tree, result of function 
\code{incidence.matrix}, actualized with coeficients computed by function 
\code{incidence_matrix_actualization_factors}.}
}
\value{
Named list, with "shifts_values" a matrix of shifts values
corresponding to the shifts positions in eq_shifts_edges; and "betas_0" a
vector of corresponding root optimal values.
}
\description{
\code{equivalent_shifts_values} computes the values of the shifts given all the 
possible allocations computed by function \code{equivalent_shifts_edges}.
}
\details{
This function uses the linear representation of the problem. It fist compute the 
mean at the tips given by the orgininal shifts positions and values, and then uses
function \code{qr.solve} (through function \code{find_actualized_shift_values}) to 
find back the values of the shifts, givent their various positions, and the means
at the tips. Function \code{compute_actualization_factors} is used to compute the
actualization factor that multipies the shifts values at the tips. Carefull, only 
work for ultrametric trees.
}

