% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{estimation_wrapper.OUsr}
\alias{estimation_wrapper.OUsr}
\title{A wrapper for estimateEM for OU with stationary root.}
\usage{
estimation_wrapper.OUsr(K_t, phylo, Y_data, alpha_known = FALSE, alpha = 0,
  method.init = "lasso", method.init.alpha = "estimation",
  method.init.alpha.estimation = c("regression", "regression.MM", "median"),
  Nbr_It_Max = 1000, tol_h_l = 10^(-2), ...)
}
\arguments{
\item{K_t}{the number of shifts allowed}

\item{phylo}{a phylogenetic tree}

\item{Y_data}{vector of data at the tips}

\item{alpha_known}{a boolean}

\item{alpha}{the value of alpha if known}

\item{method.init.alpha}{the initialization method for alpha}

\item{...}{other arguments to pass to \code{estimateEM}.}
}
\value{
summary a data frame with columns:
   - alpha_estim the estimated selection strength
   - gamma_estim the estimated root variance
   - beta_0_estim the estimated value of root optumum
   - EM_steps number of iterations needed before convergence
   - DV_estim has the EM diverged ?
   - CV_estim has the EM converged ?
   - log_likelihood log likelihood of the data using the estimated parameters
   - mahalanobis_distance_data_mean the mahalanobis distance between the data
and the estimated means at the tips
   - least_squares the mahalanobis distance, renormalized by gamma^2: 
mahalanobis_distance_data_mean * gamma_estim.
   - mean_number_new_shifts the mean number of shifts that changed over the 
iterations of the EM
   - number_equivalent_solutions the number of equivalent solutions to 
the solution found.
   - K_try the number of shifts allowed.
   - complexity the complexity for K_try
   - time the CPU time needed.

params a list of infered parameters

params_init a list of initial parameters

Zhat the reconstructed node states

edge.quality the quality of each selected edge

raw_results complete result of \code{estimateEM}
}
\description{
\code{estimation_wrapper.OUsr} call estimateEM with a set of parameters well
suited for the OU with stationary root.
It is used in \code{estimateEM_several_K.OUsr}.
}

