% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract.enumerate_parsimony}
\alias{extract.enumerate_parsimony}
\title{Extract the result of \code{enumerate_parsimony} at a node.}
\usage{
extract.enumerate_parsimony(Reconstructions,
  node = attr(Reconstructions$allocations, "ntaxa") + 1)
}
\arguments{
\item{allocations}{a list of list of matrices, result of function 
\code{enumerate_parsimony}}
}
\value{
A matrix with ntaxa + nNode columns, and as many rows as the number of
possible parsimonious reconstructions.
}
\description{
\code{extract.enumerate_parsimony} returns a matrix contianing all the possible
regime allocations for the nodes of a given subtree.
}

