% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract.parsimonyCost}
\alias{extract.parsimonyCost}
\title{Extraction of the actual number of solutions.}
\usage{
\method{extract}{parsimonyCost}(x, node = attr(x, "ntaxa") + 1, ...)
}
\arguments{
\item{x}{an object of class "\code{parsimonyCost}", result of function
\code{\link{parsimonyCost}}.}

\item{node}{the root node of the subtree. By default, the root of the tree.}

\item{...}{unused}
}
\value{
An integer giving the minimum cost of the subtree.
}
\description{
\code{extract.parsimonyCost} takes an object of class "\code{parsimonyCost}",
result of function \code{\link{parsimonyCost}}, and computes the minimum cost
at the given node.
}
\seealso{
\code{\link{parsimonyCost}}
}
