% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract.parsimonyNumber}
\alias{extract.parsimonyNumber}
\title{Extraction of the actual number of solutions.}
\usage{
extract.parsimonyNumber(Reconstructions,
  node = attr(Reconstructions$nbrReconstructions, "ntaxa") + 1)
}
\arguments{
\item{Reconstructions}{two (ntaxa + nNodes)x(nclus) matrices, result of function
\code{pasimonyNumber}, with :
   -- nbrReconstructions the matrix of number of local parsimonious solutions.
   -- costReconstructions the matrix of costs (i.e number of shifts).}

\item{node}{the root node of the subtree. By default, the root of the tree.}
}
\value{
An integer giving the number of equivalent parsimonious solutions.
}
\description{
\code{extract.parsimonyNumber} takes the two matrices computed by 
\code{parsimonyNumber}, and compute the atual number of parsimonious solution for
any subtree starting from a given node.
}
\details{
The parsimonious solutions are the one with the minimum number of shifts (that are
given by matrice costReconstructions). This function sums the number of 
solutions (given in matrice nbrReconstructions) that have the minimum number of 
shifts.
}

