% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{extract.partitionsNumber}
\alias{extract.partitionsNumber}
\title{Extract from object \code{partitionsNumber}}
\usage{
\method{extract}{partitionsNumber}(
  x,
  node = attr(x, "ntaxa") + 1,
  npart = attr(x, "npart"),
  marked = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{partitionsNumber}, result of function
\code{\link{partitionsNumber}}.}

\item{node}{the root node of the subtree where to get the result.
Default to the root of the tree.}

\item{npart}{the number of partitions (colors) allowed at the tips.
Default to the value used in the call of function
\code{\link{partitionsNumber}} (the maximum).}

\item{marked}{whether to extract the marked (TRUE) or un-marked (FALSE)
partitions. The number of models is the number of un-marked partitions.
Default to FALSE.}

\item{...}{unused.}
}
\value{
the number of partitions with npart colors, on the sub-tree starting
at node, marked or not.
}
\description{
\code{extract.partitionsNumber} extracts the number of partitions for a 
given sub-tree, either marked or non-marked.
}
\seealso{
\code{\link{partitionsNumber}}
}
