% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{extract.simul_process}
\alias{extract.simul_process}
\title{Extraction of simulated traits}
\usage{
\method{extract}{simul_process}(
  x,
  where = c("tips", "nodes"),
  what = c("states", "expectations", "optimal.values"),
  ...
)
}
\arguments{
\item{x}{an object of class "\code{simul_process}", result of function
\code{\link{simul_process}}.}

\item{where}{one of "tips" (the default) or "nodes". Where to extract the results.}

\item{what}{one of "states" (the default), "expectation", or "optimal.values".}

\item{...}{unused}
}
\value{
A matrix giving the selected quantities at the selected nodes or tips. If
the tips or nods are labeled, then the colnames of the matrix are set accordingly.
}
\description{
\code{extract.simul_process} takes an object of class "\code{simul_process}",
result of function \code{\link{simul_process}}, and extracts the traits values, 
expectations or optimal values at the tips or the internal nodes.
}
\details{
##
}
\seealso{
\code{\link{simul_process}}
}
