% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract_root_values}
\alias{extract_root_values}
\title{Extract the shifts values for one trait.}
\usage{
extract_root_values(eq_shifts, trait)
}
\arguments{
\item{eq_shifts}{an object of class \code{equivalent_shifts}, result of
function \code{\link{equivalent_shifts}}}

\item{trait}{the number of the trait to be extracted.}
}
\value{
a matrix with the values of the root on the trait for each
equivalent configuration. Each column is one configuration. Rows are the
values of the shifts given by \code{eq_shifts_edges}, the result of function
\code{\link{equivalent_shifts_edges}}.
}
\description{
\code{extract_root_values} takes an object of class
\code{equivalent_shifts}, result of function \code{\link{equivalent_shifts}},
and returns the root values at a given trait.
}
\seealso{
\code{\link{plot.equivalent_shifts}},
\code{\link{equivalent_shifts_edges}}
}

