% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{find_rotation}
\alias{find_rotation}
\title{Test for rotation invarient datasets}
\usage{
find_rotation(res1, res2, tol = NULL)
}
\arguments{
\item{res1}{an object of class \code{\link{PhyloEM}}.}

\item{res2}{an object of class \code{\link{PhyloEM}}.}

\item{tol}{relative numerical tolerence. Default to \code{.Machine$double.eps^(0.5)}.}
}
\value{
If appropriate, the rotation matrix rot such that dat1 = rot %*% dat2.
}
\description{
\code{find_rotation} takes two fits from from \code{\link{PhyloEM}},
and test if their datasets are equal up to a rotation.
}
