% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{find_shift_values}
\alias{find_shift_values}
\title{Find values given edges. OU stationary case. Ultrametric tree.}
\usage{
find_shift_values(shifts_edges, T_tree_ac, vect_Y, p, ntaxa)
}
\arguments{
\item{shifts_edges}{a vector of positions of shifts on the tree.}

\item{T_tree_ac}{matrix of incidence of the tree, result of function 
\code{incidence.matrix}.}

\item{m_Y}{the vector of values of the means at the tips.}
}
\value{
vector, with first entry the values at the root, and other entries the
values of the shifts.
}
\description{
\code{find_actualized_shift_values} computes the values of the shifts their
positions, and the mean values at the tips.
Warning : this function does not check for consistency. Please make sure that the
shifts positions and the mean values are compatible.
}
\details{
This function uses \code{qr.solve} for rectangular linear system solving.
}
\keyword{internal}
