% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get_criterion}
\alias{get_criterion}
\title{Get Model Selection Criterion}
\usage{
get_criterion(res, method.selection = NULL)
}
\arguments{
\item{res}{an object of class \code{PhyloEM}, result of function
\code{\link{PhyloEM}}.}

\item{method.selection}{select the parameters to plot. One of "LINselect", "DDSE",
"Djump" or "likelihood" (for un-penalized likelihood). Default to "LINselect". See
\code{\link{params_process.PhyloEM}}.}
}
\value{
A named vector with the values of the criterion for each number of shift K.
}
\description{
This function takes an object of class \code{PhyloEM}, result of function
\code{\link{PhyloEM}}, and return the values of the model selection criterion
for each value of K.
}
\seealso{
\code{\link{params_process.PhyloEM}}, \code{\link{plot.PhyloEM}}, \code{\link{plot_criterion}}
}
