% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{go_back_to_original_process}
\alias{go_back_to_original_process}
\title{Scale the parameters back to the original process}
\usage{
go_back_to_original_process(
  phy_original,
  known.selection.strength,
  sBM_variance,
  params
)
}
\arguments{
\item{phy_original}{the original phylogenetic tree}

\item{known.selection.strength}{the known selection strength of the original
OU.}

\item{sBM_variance}{boolean. Is the root random ?}

\item{params}{the inferred parameters of the BM on the re-scaled tree.}
}
\value{
params_scOU the equivalent parameters of the OU on the original tree.
}
\description{
\code{go_back_to_original_process} takes the inferred parameters with a BM
on a rescaled tree, and gives back the equivalent parameters of the OU on 
the original process.
}
\keyword{internal}
