% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_EM.R
\name{impute.data.Rphylopars}
\alias{impute.data.Rphylopars}
\title{Initial imputation of missing data for lasso}
\usage{
impute.data.Rphylopars(phylo, Y_data, process, random.init)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}

\item{Y_data}{data at the tips.}

\item{process}{the stochastic process}

\item{random.init}{whether root is random or fixed.}
}
\value{
Y_data_imp the imputed data using Rphylopars
}
\description{
\code{impute.data.Rphylopars} uses function \code{phylopars} from package \code{Rphylopars}
to impute missing data.
}
\details{
This function assume that there are no shifts on the tree. It is only a first approximation
for initialization purposes.
}
\keyword{internal}
