% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{imputed_traits.PhyloEM}
\alias{imputed_traits.PhyloEM}
\title{Ancestral State Reconstruction}
\usage{
imputed_traits.PhyloEM(x, trait = 1, where = c("nodes", "tips"),
  method.selection = method.selection)
}
\arguments{
\item{x}{an object of class \code{\link{PhyloEM}}.}

\item{trait}{an integer giving the trait to extract. Default to 1.}

\item{where}{either "nodes" for ancestral state reconstruction, or "tips" for
data imputation.}

\item{method.selection}{(optional) the method selection to be used.
One of "BGH", "DDSE", "Djump". Default to "BGH" for univariate, and DDSE for
multivariate.}
}
\value{
A vector with ancestral trait value at the nodes of the tree.
}
\description{
\code{imputed_traits.PhyloEM} takes an object of class \code{\link{PhyloEM}},
and returns the imputed traits values, either at the internal nodes (ancestral
state reconstruction) or at the tips (data imputation)
}

