% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{incidence.matrix.full}
\alias{incidence.matrix.full}
\title{Incidence matrix of a tree.}
\usage{
incidence.matrix.full(phylo)
}
\arguments{
\item{phylo}{Input tree.}
}
\value{
Matrix of incidence, size ntaxa + nNodes
}
\description{
\code{incidence.matrix.full} computes the incidence matrix T of a tree : for a 
node i and a branch b, T[i,b]=1 if b is in the lineage i, and 0 
otherwise.
}
\details{
This function uses the general up tree recursion function \code{recursionUp}
with the initialization function \code{init.incidence.matrix.full} and update
function \code{update.incidence.matrix.full}.
}

