% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{init.allocate_regimes_from_shifts}
\alias{init.allocate_regimes_from_shifts}
\title{Initialisation for the allocation of shifts.}
\usage{
init.allocate_regimes_from_shifts(phy, ...)
}
\arguments{
\item{phy}{Input tree.}
}
\value{
Matrix of size (nNodes + ntaxa)x1 of NAs, with the 0 at the root.
}
\description{
\code{init.allocate_regimes_from_shifts} initialize the vector of regimes 
at nodes and tips, with the regime (0) at the root.
}
\details{
This function is used in function \code{allocate_regimes_from_shifts} and is 
designed to furnish function \code{update.allocate_regimes_from_shifts} with 
the right structure of data.
}
\keyword{internal}
