% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{init.compute_betas}
\alias{init.compute_betas}
\title{Initialisation for the computation of the optimal values}
\usage{
init.compute_betas(phy, optimal.value, ...)
}
\arguments{
\item{phy}{Input tree.}

\item{optimal.value}{the optimal value at the root of the tree}
}
\value{
Matrix of size (nNodes + ntaxa)x1 of NAs, with the optiaml value
 at the root.

06/10/14 - Initial release
}
\description{
\code{init.compute_betas} initialize the vector of optimal values at nodes and
tips, with the value at the root.
}
\details{
This function is used in function \code{compute_betas} and is designed to 
furnish function \code{update.compute_betas} with the right structure of data.
}

