% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{init.incidence.matrix.full}
\alias{init.incidence.matrix.full}
\title{Initialization for incidence matrix (full tree)}
\usage{
init.incidence.matrix.full(phy)
}
\arguments{
\item{phy}{Input tree.}
}
\value{
Matrix of size ntaxa + nNodes.
}
\description{
\code{init.incidence.matrix.full} initialize the matrix updated in
\code{update.incidence.matrix.full} for the computation of the incidence
 matrix of the full tree in \code{incidence.matrix.full}.
}
\details{
The initialized matrix is squared of size ntaxa + nNodes. Each node
represent its parental branch. A row corresponding to a tip i is initialized
to a vector of zeros, with only entry i equal to one. (Branch ending at 
tip i is only in the i^th lineage)
}
\keyword{internal}
