% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{init.parsimonyCost}
\alias{init.parsimonyCost}
\title{Initialization for parsimonyCost.}
\usage{
init.parsimonyCost(phy, clusters)
}
\arguments{
\item{clusters}{the vector of the clusters of the tips.}

\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}
}
\value{
A (ntaxa + nNodes)x(nclus) matrix, with ntaxa first lines initialized as
described.
}
\description{
\code{init.parsimonyCost} initialize a (ntaxa + nNodes) x (nclus) matrix with
NAs everywhere, except for the tips.
}
\details{
At a tip i in state k, the line-vector is initialized as follow : 
(1 - Ind(k=p)_{1<=p<=nclus})*Inf (where Inf * 0 = 0)
}
\keyword{internal}
