% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{init.parsimonyNumber}
\alias{init.parsimonyNumber}
\title{Initialization for parsimonyNumber.}
\usage{
init.parsimonyNumber(phy, clusters)
}
\arguments{
\item{phy}{phylogenetic tree.}

\item{clusters}{the vector of the clusters of the tips.}
}
\value{
A (ntaxa + nNodes)x(nclus) matrix, with ntaxa first lines initialized
as described.
}
\description{
\code{init.parsimonyNumber} initialize a (ntaxa + nNodes)x(nclus) matrix with
NAs everywhere, except for the tips.
}
\details{
At a tip i in state k, the line-vector is initialized as follow : Ind(k=p)_{1<=p<=nclus}
}
\keyword{internal}

