% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutoff.R
\name{is.finite.params.BM}
\alias{is.finite.params.BM}
\title{Check wether parameters are finite.}
\usage{
\method{is.finite}{params.BM}(params)
}
\arguments{
\item{params}{list of parameters with the correct structure}
}
\value{
boolean

10/06/14 - Initial release
}
\description{
\code{is.finite.params} checks whether calculated parameters in the EM are
finite or not.
}
\details{
This function is used to test the convergence of the algorithm.
}

