% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{log_likelihood.OU}
\alias{log_likelihood.OU}
\title{Log Likelihood of a model}
\usage{
log_likelihood.OU(Y, phylo, params, ...)
}
\arguments{
\item{Y}{the vector of the data at the tips}

\item{phylo}{a phylogenetic tree}

\item{params}{list of parameters with the correct structure}
}
\value{
boolean
}
\description{
\code{likelihood.OU} computes the likelihhod of the data given a model. This
is a non-efficient debugging function.
}
\details{
This function uses functions \code{compute_mean_variance.simple}, \code{compute_times_ca}, \code{compute_dist_phy}, \code{compute_log_likelihood.simple}
}
\keyword{internal}

