% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{matrix_of_possibles}
\alias{matrix_of_possibles}
\title{Compute parent matrix from possibles daughter matrices.}
\usage{
matrix_of_possibles(matrices)
}
\arguments{
\item{matrices}{a list of matrices with ndaughters entries.}
}
\value{
Matrix of all possible regimes fot the subtree bellow node parent.
}
\description{
\code{matrix_of_possibles} is used in function \code{update.enumerate_parsimony}
to compute, from the list of possible matrices for the daughters, the matrix for
the node (a group for the parent being fixed).
}
\details{
This function select all possible combinations of rows from all the daughters, and
merge then into one using function \code{merge_complementary_vectors}.
}
\keyword{internal}
