% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{merge_complementary_vectors}
\alias{merge_complementary_vectors}
\title{Merge several complementary vectors into one.}
\usage{
merge_complementary_vectors(comb, mat)
}
\arguments{
\item{comb}{vector giving the rows to be kept.}

\item{mat}{matrice containing the vectors as rows.}
}
\value{
row vector of the same size as entry matrix.
}
\description{
\code{merge_complementary_vectors} take several vectors with complementary entries
(i.e, vector of same length, that are such that only one vector has a non NA value
for each entry), and merge them into one.
}
\details{
The vectors are selected from the entry matrix by comb. At each entry, the vectors
are added using function \code{add_complementary}.
}
\keyword{internal}

