% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{merge_params_independent}
\alias{merge_params_independent}
\title{Merge a list of independent parameters into into one parameter}
\usage{
merge_params_independent(params_split)
}
\arguments{
\item{params_split:}{a list of parameters}
}
\value{
A parameter object
}
\description{
\code{merge_params_independent} merges a list of p params
objects into one param object of dimension p
The reverse operation is done by \code{split_params_independent}
}
\keyword{internal}

