% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{params.PhyloEM}
\alias{params.PhyloEM}
\title{Parameter estimates}
\usage{
params.PhyloEM(x, method.selection)
}
\arguments{
\item{x}{an object of class \code{\link{PhyloEM}}}

\item{method.selection}{(optional) the method selection to be used.
One of "BGH", "DDSE", "Djump". Default to "BGH" for univariate, and DDSE for
multivariate.}
}
\value{
An object of class \code{\link{params_process}}.
}
\description{
\code{params} takes an object of class \code{\link{PhyloEM}}, and returns the 
infered parameters of the process.
}

