% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_EM.R
\name{params_process.character}
\alias{params_process.character}
\title{Create an object \code{params_process}}
\usage{
\method{params_process}{character}(x, ...)
}
\arguments{
\item{x}{one of "BM" or "OU"}

\item{...}{specified parameters, see functions \code{\link{params_BM}} and 
\code{\link{params_OU}} for details.}
}
\value{
an object of class \code{params_process}.
}
\description{
\code{params_process} creates a coherent object params_process from user 
provided values of the paramerters.
}
\seealso{
\code{\link{params_BM}}, \code{\link{params_OU}}
}

