% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_EM.R
\name{params_process}
\alias{params_process}
\title{Create an object params_process}
\usage{
params_process(x, ...)
}
\arguments{
\item{x}{an S3 object.}

\item{...}{further arguments to be passed to the specific method.}
}
\value{
An S3 object of class \code{params_process}.
}
\description{
\code{params_process} creates or extracts a set of parameters of class
\code{params_process}.
}
\seealso{
\code{\link{params_process.character}},
\code{\link{params_process.PhyloEM}},
\code{\link{params_BM}}, \code{\link{params_OU}}
}

