% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{parsimonyCost}
\alias{parsimonyCost}
\title{Minimal number of shifts needed to get a clustering.}
\usage{
parsimonyCost(phylo, clusters = rep(1, length(phylo$tip.label)))
}
\arguments{
\item{phylo}{phylogenetic tree.}

\item{clusters}{the vector of the clusters of the tips.}
}
\value{
A (ntaxa + nNodes) x (nclus) matrix of the total number of shifts needed 
to get the clustering, if starting from a node in state k.
}
\description{
\code{parsimonyCost} is an implementation of the Sankoff algorithm, when the cost of
transition between two state is always one.
}
\details{
This functin does a recursion up the tree, using functions 
\code{init.parsimonyCost} for the initialization at the tips, 
\code{updateUp} for the actual recursion on the tree,
and \code{update.parsimonyCost} for the actualisation of the parameters.
}

