% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{parsimonyNumber}
\alias{parsimonyNumber}
\title{Number of equivalent parsimonious allocations.}
\usage{
parsimonyNumber(phylo, clusters = rep(1, length(phylo$tip.label)))
}
\arguments{
\item{phylo}{phylogenetic tree.}

\item{clusters}{the vector of the clusters of the tips.}
}
\value{
an object of S3 class "\code{parsimonyNumber}" with:
\describe{
 \item{nbrReconstructions}{a (ntaxa + nNodes) x (nclus)
 matrix of loccaly parsimonious solutions starting from a cluster k at a
 given node}
 \item{costReconstructions}{an object of class "\code{parsimonyCost}",
 result of function \code{\link{parsimonyCost}}.}
}
}
\description{
\code{parsimonyNumber} aims at finding the number of equivalent allocations of
the shifts on the tree, i.e allocations that are parsimonious and compatible
with a given clustering of the tips.
}
\details{
This function does a recursion up the tree.
The function \code{\link{extract.parsimonyNumber}} gives the result seeked for
any subtree.
The matrix of costs of the states (number of shifts) is also required, it is
computed by function \code{\link{parsimonyCost}}.
}
\seealso{
\code{\link{extract.parsimonyNumber}},
\code{\link{partitionsNumber}}, \code{\link{parsimonyCost}}
}

