% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{partitionsNumber}
\alias{partitionsNumber}
\title{Number of different models}
\usage{
partitionsNumber(phylo, npart)
}
\arguments{
\item{phylo}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}

\item{npart}{the numbers of partitions (colors) allowed at the tips. This
is the number of shifts plus one.}
}
\value{
an object of class \code{partitionsNumber}. This is made of a matrix
with Nnodes+ntaxa rows and 2*npart columns. Each column contains two vectors :
for k=1:npart it contains the number of partitions with k groups compatible
with the tree and the shift process; and for k=(npart+1):2*npart, it contains
the number of "marqued" partitions with (k-npart) groups compatible with the
tree and the shift process.
}
\description{
\code{partitionsNumber} computes the number of different models with a given
number of shifts K. It is also the number of colorings of the tips to the 
tree in npart=K+1 colors.
}
\references{
Paul Bastide, Mahendra Mariadassou, Stéphane Robin:
Detection of adaptive shifts on phylogenies using shifted stochastic processes
on a tree.
}
\seealso{
\code{\link{extract.partitionsNumber}}, \code{\link{parsimonyNumber}}
}

