% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{penalty_pBIC}
\alias{penalty_pBIC}
\title{Penalty function type pBIC}
\usage{
penalty_pBIC(all_params, model_complexity, independent, tree, times_shared,
  distances_phylo, T_tree, p, K, ntaxa, process)
}
\arguments{
\item{model_complexity}{the complexity of the set of models with dimension K.}

\item{K}{the dimension of the model.}

\item{ntaxa}{the number of tips.}

\item{C}{a constant, C > 1. Default is C = 1.1
(as suggested in Baraud Giraud Huet (2009))}
}
\value{
value of the penalty.
}
\description{
\code{penalty_pBIC_scalarOU} is the pBIC.
}
\seealso{
\code{\link{penalty_BirgeMassart_shape1}},
\code{\link{penalty_BirgeMassart_shape2}}
}
\keyword{internal}
