% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{PhyloEM}
\alias{PhyloEM}
\title{Phylogenetic EM}
\usage{
PhyloEM(phylo, Y_data, process = c("BM", "OU", "scOU", "rBM"),
  check_postorder = TRUE, independent = FALSE, K_max, use_previous = TRUE,
  order = TRUE, method.selection = c("BirgeMassart1", "BirgeMassart2",
  "BGH"), C.BM1 = 0.1, C.BM2 = 2.5, C.BGH = 1.1,
  method.variance = c("simple", "upward_downward"), method.init = "lasso",
  method.init.alpha = "estimation",
  method.init.alpha.estimation = c("regression", "regression.MM", "median"),
  methods.segmentation = c("lasso", "best_single_move"), alpha_grid = TRUE,
  random.root = FALSE, stationary.root = FALSE, alpha = NULL,
  check.tips.names = FALSE, progress.bar = TRUE, estimates = NULL,
  save_step = FALSE, sBM_variance = FALSE, method.OUsun = "rescale",
  parallel_alpha = FALSE, Ncores = 3, exportFunctions = ls(),
  impute_init_Rphylopars = FALSE, K_lag_init = 0, ...)
}
\arguments{
\item{phylo}{a phylogenetic tree}
}
\value{

}
\description{

}

