% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{plot.enumerate_parsimony}
\alias{plot.enumerate_parsimony}
\title{Plot all the equivalent solutions.}
\usage{
\method{plot}{enumerate_parsimony}(x, numbering = FALSE, nbr_col = 3,
  ...)
}
\arguments{
\item{x}{an object of class \code{enumerate_parsimony}, result of
function \code{\link{enumerate_parsimony}}}

\item{numbering}{whether to number the solutions. Default to FALSE.}

\item{nbr_col}{the number of columns on which to display the plot.
Default to 3.}

\item{...}{further arguments to be passed to \code{\link[ape]{plot.phylo}} or 
\code{\link[ape]{nodelabels}}.}
}
\value{
A plot of the equivalent shifts allocations.
}
\description{
\code{plot.enumerate_parsimony} plots a representation of all the equivalent
solutions.
}
\details{
This function uses functions \code{\link[ape]{plot.phylo}} and
\code{\link[ape]{nodelabels}} for the actual plotting of the trees.
}
\seealso{
\code{\link[ape]{plot.phylo}}, \code{\link{enumerate_parsimony}},
\code{\link{plot.equivalent_shifts}}, \code{\link[ape]{nodelabels}}
}
