% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_criterion}
\alias{plot_criterion}
\title{Plot Model Selection Criterion}
\usage{
plot_criterion(
  res,
  method.selection = NULL,
  add = FALSE,
  select.col = "red",
  ...
)
}
\arguments{
\item{res}{an object of class \code{PhyloEM}, result of function
\code{\link{PhyloEM}}.}

\item{method.selection}{select the parameters to plot. One of "LINselect", "DDSE",
"Djump" or "likelihood" (for un-penalized likelihood). Default to "LINselect". See
\code{\link{params_process.PhyloEM}}.}

\item{add}{boolean: should the points be added to a current plot (default to FALSE).}

\item{select.col}{the color of the point selected by the criterion. Default to "red".}

\item{...}{further argument to be passed to base \code{\link{plot}}.}
}
\description{
This function takes an object of class \code{PhyloEM}, result of function
\code{\link{PhyloEM}}, and plots a model selection criterion.
}
\seealso{
\code{\link{params_process.PhyloEM}}, \code{\link{plot.PhyloEM}}, \code{\link{get_criterion}}
}
