% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{plot_equivalent_shifts}
\alias{plot_equivalent_shifts}
\title{Plot all the equivalent solutions.}
\usage{
plot_equivalent_shifts(phylo, eq_shifts_edges, eq_shifts_values, PATH, name)
}
\arguments{
\item{phylo}{a phylogenetic tree.}

\item{eq_shifts_edges}{matrix, result of function \code{equivalent_shifts_edges}.}

\item{eq_shifts_values, }{result of function \code{equivalent_shifts_values}.}

\item{PATH}{a string linking to the path were the plot is to be stored.}

\item{name}{a name to be given to the plot.}
}
\description{
\code{plot_equivalent_shifts} plots and save a representation of all the equivalent
shifts allocations, with a representation of the shifts and their values, and a 
coloration of the branches in term of regimes. Black is the ancestral regime.
}
\details{
This function uses \code{qr.solve} for rectangular linear system solving.
}

