% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{print.enumerate_parsimony}
\alias{print.enumerate_parsimony}
\title{Display the number of parsimonious solutions}
\usage{
\method{print}{enumerate_parsimony}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{enumerate_parsimony}}.}

\item{...}{unused}
}
\description{
\code{print.enumerate_parsimony} prints the number of equivalent parsimonious
allocations of shifts, from the root of the tree, using function
\code{\link{extract.enumerate_parsimony}}.
}
\seealso{
\code{\link{enumerate_parsimony}},
\code{\link{extract.enumerate_parsimony}},
\code{\link{plot.enumerate_parsimony}}
}

