% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{print.parsimonyCost}
\alias{print.parsimonyCost}
\title{Display a parsimony cost}
\usage{
\method{print}{parsimonyCost}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{parsimonyCost}}.}

\item{...}{unused}
}
\description{
\code{print.parsimonyCost} prints the parsimony cost at the root of the tree,
using function \code{\link{extract.parsimonyCost}}.
}
\seealso{
\code{\link{parsimonyCost}}, \code{\link{extract.parsimonyCost}}
}

