% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{print.parsimonyNumber}
\alias{print.parsimonyNumber}
\title{Display the number of parsimonious solutions}
\usage{
\method{print}{parsimonyNumber}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{parsimonyNumber}}.}

\item{...}{unused}
}
\description{
\code{print.parsimonyNumber} prints the number of equivalent parsimonious
allocations of shifts, from the root of the tree, using function
\code{\link{extract.parsimonyNumber}}.
}
\seealso{
\code{\link{parsimonyNumber}}, \code{\link{extract.parsimonyNumber}}
}

