% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{print.partitionsNumber}
\alias{print.partitionsNumber}
\title{Display the number of models}
\usage{
\method{print}{partitionsNumber}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{partitionsNumber}}.}

\item{...}{unused}
}
\description{
\code{print.partitionsNumber} prints the number of tree-compatible colorings of
the tips of the tree, using functio \code{\link{extract.partitionsNumber}}.
}
\seealso{
\code{\link{partitionsNumber}}, \code{\link{extract.partitionsNumber}}
}

