% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{prod.index}
\alias{prod.index}
\title{Product of elements of a matrix}
\usage{
\method{prod}{index}(X, Id)
}
\arguments{
\item{X}{a martix with p rows and K column. Each row contains the number of
partition in 1<=k<=K groups for one of the p children of a given node}

\item{Id}{a vector of length p, result of the function \code{xsimplex}.}
}
\value{
double : the result of the product.
}
\description{
\code{prod.index} return the procuct of choosen elements of a matrix.
}
\details{
This function is to be used in \code{sum.simplex} to be applied to all the
elements given by \code{xsimplex}.
Performs the product : X[1,Id[1]]*X[2,Id[2]]*...*X[p,Id[p]] if all the
elements of Id are positive. Otherwise, just return 0.
}
\keyword{internal}
