% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{qr_solve_exact}
\alias{qr_solve_exact}
\title{exact qr.solve}
\usage{
qr_solve_exact(a, b, tol = 1e-07)
}
\arguments{
\item{a}{a QR decomposition or a rectangular matrix.}

\item{b}{a vector or matrix of right-hand sides of equations.
\code{incidence.matrix}.}

\item{tol}{the tolerance for detecting linear dependencies in the columns of x.
Only used if LAPACK is false and x is real.}
}
\description{
This is the same function as \code{qr.solve}, but it throws an error if an exact fit cannot
be found (instead of returning a least square fitted value, which is the default behavior
of \code{qr.solve}).
}
\keyword{internal}
