% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{recursionDown}
\alias{recursionDown}
\title{Generic recursion down the tree.}
\usage{
recursionDown(phy, params, updateDown, ...)
}
\arguments{
\item{phy}{Input tree, in cladewise order.}

\item{params}{Matrix of parameters to update by the recursion}

\item{updateDown}{Function to be used for the update}

\item{...}{Arguments to be used by the function updateDown}
}
\value{
Matrix of parameters updated.
}
\description{
\code{recursionDown} uses the function \code{updateDown} to compute
daughters rows of matrix param.
}
\details{
This function is to be used in other more complex function that need to
update a quantity from the root to the tips of a tree. Note that the
input tree must be in cladewise order.
}
\keyword{internal}
