% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{residuals.PhyloEM}
\alias{residuals.PhyloEM}
\title{Residuals of a fitted object}
\usage{
\method{residuals}{PhyloEM}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{params_process}} or \code{\link{PhyloEM}}.
trait, and each column is a tip. The column names are checked against the
tip names of the tree.
\code{\link{incidence.matrix.full}}. Can be precised to avoid extra computations.}

\item{...}{for a \code{PhyloEM} object, further arguments to be passed on to
\code{\link{params_process.PhyloEM}} (to choose which parameters to extract from
the results, see documentation of this function).}
}
\value{
The log likelihood of the data with the provided parameters on the tree.
}
\description{
\code{residuals} computes the residuals of some parameters.
}
\seealso{
\code{\link{params_process}}, \code{\link{PhyloEM}}
}

