% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{sample_shifts_edges}
\alias{sample_shifts_edges}
\title{Sample shifts edges in a parsimonious way.}
\usage{
sample_shifts_edges(tree, K,
  part.list = enumerate_tips_under_edges(tree))
}
\arguments{
\item{tree}{: input tree}

\item{K}{: number of edges to be sampled.}
}
\value{
vector of edges
}
\description{
\code{sample_shifts_edges} attempts to find K shifts in the tree which allocations
are parsimonious. The actual generation of edges is done by function 
\code{sample_edges_intervals}.
}
\details{
This function uses function \code{enumerate_tips_under_edges} to generate a list 
of tips under each edge, and function \code{check_parsimony} to check for
parsimony of a given solution, under the assumption of an "infinite site model".
}
\keyword{internal}
