% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{scale_params}
\alias{scale_params}
\title{Scale variance and selection strenght from a linear transform}
\usage{
scale_params(params, f)
}
\arguments{
\item{params}{Parameters list}

\item{f}{Factor of the linear transform. If t' = f * t, the function takes
parameters from phylo' back to phylo.}
}
\value{
re-scaled parameters
}
\description{
Used for process equivalencies on re-scaled trees.
}
\keyword{internal}
