% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{segmentation.BM}
\alias{segmentation.BM}
\title{Segmentation in the BM case}
\usage{
segmentation.BM(nbr_of_shifts, costs0, diff_exp)
}
\arguments{
\item{nbr_of_shifts}{Number of shifts on the phylogeny allowed}

\item{costs0}{Cost of each edge}

\item{diff_exp}{Difference of expectations}
}
\value{
List containing : edges_max : array of nbr_of_shifts edges where
costs0 is maximal.
                          shifts:list containing the computed tau and delta
}
\description{
\code{segmentation.BM} performs the segmentation algorithm described.
}
\details{
This function takes the largest values of costs0, and make them null,
 thanks to delta and tau.
}
\keyword{internal}

