% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{segmentation.OU.specialCase.max_costs_0}
\alias{segmentation.OU.specialCase.max_costs_0}
\title{Segmentation in the OU special case, algo 1}
\usage{
segmentation.OU.specialCase.max_costs_0(phylo, nbr_of_shifts, conditional_law_X,
  selection.strength, ...)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{nbr_of_shifts}{Number of shifts on the phylogeny allowed}

\item{conditional_law_X}{moments of the conditional law of X given Y, result
of function \code{compute_M.OU.specialCase}}

\item{selection.strength}{the selection strength}
}
\value{
List containing : beta_0 : the optimal value at the root
                          shifts : list containing the computed tau and delta
                          costs : vector of costs
}
\description{
\code{segmentation.OU.specialCase.max_costs_0} performs the first segmentation 
algorithm described.
}
\details{
This function takes the largest values of costs0, and make them null,
 thanks to delta and tau.
}
\keyword{internal}

