% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{segmentation.OU.specialCase.same_shifts}
\alias{segmentation.OU.specialCase.same_shifts}
\title{Segmentation in the OU special case, conserving the same shifts
 position.}
\usage{
segmentation.OU.specialCase.same_shifts(phylo, shifts_old, D, Xp, ...)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{shifts_old}{the previous list of shifts (only position is used)}
}
\value{
List containing : beta_0 : the optimal value at the root
                          shifts : list containing the computed tau and delta
                          costs : vector of costs
}
\description{
\code{segmentation.OU.specialCase.same_shifts} keeps the same shifts position,
and optimize the sum of costs using function 
\code{best_scenario}
}
\details{
This is the best move if keeping the previous shifts positions.
}
\keyword{internal}
