% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_step.R
\name{segmentation.OU.specialCase.same_shifts_same_values}
\alias{segmentation.OU.specialCase.same_shifts_same_values}
\title{Segmentation in the OU special case, conserving the same shifts.}
\usage{
segmentation.OU.specialCase.same_shifts_same_values(phylo, conditional_law_X,
  selection.strength, beta_0_old, shifts_old, ...)
}
\arguments{
\item{phylo}{a phylogenetic tree}

\item{conditional_law_X}{moments of the conditional law of X given Y, result
of function \code{compute_M.OU.specialCase}}

\item{selection.strength}{the selection strength}

\item{beta_0_old}{the previous (and concerved) value of beta_0}

\item{shifts_old}{the previous (and concerved) list of shifts}
}
\value{
List containing : beta_0 : the optimal value at the root
                          shifts : list containing the computed tau and delta
                          costs : vector of costs
}
\description{
\code{segmentation.OU.specialCase.same_shifts_same_values} keeps the same
parameters and compute the quantities needed. It is here to ensure that we do
at least better than the previous step.
}
\details{
This function takes the old shifts parameters, and compute costs with the new
 moments.
}
\keyword{internal}

