% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{shifts.list_to_matrix}
\alias{shifts.list_to_matrix}
\title{Compute the matrix of shifts.}
\usage{
shifts.list_to_matrix(phy, shifts, p = nrow(shifts$values))
}
\arguments{
\item{phy}{Input tree.}

\item{shifts}{list description of the shifts : shifts$edges, shifts$values.}

\item{p}{number of traits (optional, needed when shifts = NULL).}
}
\value{
Matrix p x nEdges of length nbranch.
}
\description{
\code{shifts.list_to_matrix} takes the list description of the shifts 
to give the matricial representation of the shifts : the b th element of 
the lth line has the value of the shift on character l occuring on that branch b
}
\seealso{
\code{\link{shifts.matrix_to_list}}
}

