% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{shifts.list_to_vector}
\alias{shifts.list_to_vector}
\title{Compute the vector of shifts.}
\usage{
shifts.list_to_vector(phy, shifts)
}
\arguments{
\item{phy}{Input tree.}

\item{shifts}{: list description of the shifts : shifts$edges, shifts$values}
}
\value{
Vector of length nbranch.
}
\description{
\code{shifts.list_to_vector} takes the list description of the shifts 
to give the vectorial representation of the shifts : the b th element of 
the vector has the value of the shift occurring on that branch b.
}
\seealso{
\code{\link{shifts.vector_to_list}}
}
\keyword{internal}
