% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{shifts.vector_to_list}
\alias{shifts.vector_to_list}
\title{Compute the list of shifts.}
\usage{
shifts.vector_to_list(delta)
}
\arguments{
\item{delta}{: vector description of the shift.}
}
\value{
Vector of length nbranch.
}
\description{
\code{shifts.vector_to_list} takes the vectorial description of the shifts 
to create the list description of the shifts.
}
\seealso{
\code{shifts.list_to_vector}
}
\keyword{internal}

