% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{shifts_to_simmap}
\alias{shifts_to_simmap}
\title{Simmap format mapping from list of edges}
\usage{
shifts_to_simmap(tree, shifts_edges)
}
\arguments{
\item{tree}{imput tree in \code{\link[ape]{phylo}} format}

\item{shifts_edges}{shifts positions on the edges}
}
\value{
tree a simmap object
}
\description{
\code{shifts_to_simmap} takes a vector of edges where the shifts occur, and return a
simmap formated tree, mapped with corresponding regimes.
}
\details{
Ancestral state is always 0, and other states are consecutive integers.
}
