% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate a Stochastic Process on a tree}
\usage{
simulate(phylo, process = c("BM", "OU", "scOU", "StudentOU"), p = 1,
  root.state = list(random = FALSE, stationary.root = FALSE, value.root,
  exp.root, var.root), shifts = list(edges = NULL, values = NULL,
  relativeTimes = NULL), eps = 10^(-6), selection.strength = NULL,
  variance = NULL, optimal.value = NULL, checks = TRUE,
  simulate_random = TRUE, U_tree = NULL, df = 1)
}
\arguments{
\item{root.state}{(list): state of the root, with:
random : random state (TRUE) or deterministic state (FALSE)
value.root : if deterministic, value of the character at the root
exp.root : if random, expectation of the character at the root
var.root : if random, variance of the character at the root (pxp matrix)}

\item{shifts}{(list) position and values of the shifts :
edges : vector of the K id of edges where the shifts are
values : matrix p x K of values of the shifts on the edges (one column = one shift)
relativeTimes : vector of dimension K of relative time of the shift from the
parentnode of edges}

\item{phylo:}{Input tree.}

\item{process:}{choose BM or OU process}

\item{p:}{dimention of the trait simulated}

\item{eps:}{tolerance for the value of the norm 1 of the selection strength matrix for OU}

\item{variance:}{variance-covariance matrix size p x p}

\item{selection.strenght:}{matrix of selection strength size p x p (OU)}

\item{optimal.value:}{vector of p optimal values at the root (OU)}
}
\value{
paramSimu: array p x nNodes x 2 (BM) or p x nNodes x 3 (OU). For each trait t, 
1 <= t <= p, paramSimu[t, , ] has tree columns, containing respectively the simulated state,
 expected value and optimal value for all the nodes.

16/05/14 - Initial release
20/05/14 - Gestion of edges (function correspondanceEdges)
21/05/14 - extraction of the recursion
16/06/14 - check.selection.strength
24/08/15 - Multivariate
}
\description{
\code{simulate} simulate a stochastic process on a tree.
}

