% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{split_params_independent}
\alias{split_params_independent}
\title{Split independent parameters into a list of parameters}
\usage{
split_params_independent(params)
}
\arguments{
\item{params}{parameters}
}
\value{
A list of p parameters
}
\description{
\code{split_params_independent} split a params object for a 
process with p independent traits into p params objects.
The reverse operation is done by \code{merge_params_independent}
}
\keyword{internal}
