% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{sum.partitions.cardFixed}
\alias{sum.partitions.cardFixed}
\title{Sum on subsets of a given cardinal.}
\usage{
\method{sum}{partitions.cardFixed}(A, N, K, p, cardI)
}
\arguments{
\item{A}{a matrix with p rows and K column. Each row contains the number of
marked partition in 1<=k<=K groups for one of the p children of a
given node.}

\item{N}{a matrix with p rows and K column. Each row contains the number of
partition in 1<=k<=K groups for one of the p children of a given node}

\item{K}{an integer. The number of groups wanted.}

\item{p}{an integer. The number of daughters of a node.}

\item{cardI}{an integer. The cardinal of the subset wanted.}
}
\value{
double : the result of the sum.
}
\description{
\code{sum.partitions.cardFixed} returns the sum on I subset of [1,p], |I|
fixed, of the sums computed by \code{sum.prod.comb}.
}
\details{
This function uses \code{combn} to enumerate all the subsets I of [1,p] of
a given cardinal, and then performs the wanted sum on these subsets.
}
\keyword{internal}
