% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{sum.partitions}
\alias{sum.partitions}
\title{Sum on all subsets.}
\usage{
\method{sum}{partitions}(A, N, K, p, m)
}
\arguments{
\item{A}{a martix with p rows and K column. Each row contains the number of
marqued partition in 1<=k<=K groups for one of the p children of a
given node.}

\item{N}{a martix with p rows and K column. Each row contains the number of
partition in 1<=k<=K groups for one of the p children of a given node}

\item{K}{an integer. Tne number of groups wanted.}

\item{p}{an integer. The number of daughters of a node.}

\item{m}{an integer. The minimum cadinal of a subset allowed.}
}
\value{
double : the result of the sum.
}
\description{
\code{sum.partitions} returns the sum on all I subset of [1,p], with |I|>m.
}
\details{
This function applies \code{sum.partitions.cardFixed} to all integer between
m and p, and sum the results.
}
\keyword{internal}

