% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{sum.prod.comb}
\alias{sum.prod.comb}
\title{Sum on a simplex}
\usage{
\method{sum}{prod.comb}(I, A, N, K, p)
}
\arguments{
\item{I}{a vector of integers representing a subset of [1,p]}

\item{A}{a matrix with p rows and K column. Each row contains the number of
marked partition in 1<=k<=K groups for one of the p children of a
given node.}

\item{N}{a matrix with p rows and K column. Each row contains the number of
partition in 1<=k<=K groups for one of the p children of a given node}

\item{K}{an integer. The number of groups wanted.}

\item{p}{an integer. The number of daughters of a node.}
}
\value{
double : the result of the sum.
}
\description{
\code{sum.prod.comb} returns the sum on k_1+...+k_p=K+|I|-1, k_i>0 of the
products of prod(A[i,k_i], i in I)*prod(N[i,k_i], i not in I).
}
\details{
This function uses \code{sum.simplex} to perform the wanted sum on a ad-hoc
matrix, combination of rows of A and N.
}
\keyword{internal}
