% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{test.root.state}
\alias{test.root.state}
\title{Test state of root.}
\usage{
test.root.state(root.state, process = c("BM", "OU", "scOU"), ...)
}
\arguments{
\item{root.state}{A list giving the root state}

\item{process}{"BM", "OU" or "scOU"}

\item{...}{parameters of the process (if OU)}
}
\value{
Coherent list root.state.
}
\description{
\code{test.root.state} test wether the parameters of root.state given
by the user are coherent. If not, it returns a new corrected list to
define root.state.
}
\details{
To test coherence, the following priorities are applied:
 random > stationary.root > values.root = exp.root = var.root
}
\keyword{internal}
