% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{transform_branch_length}
\alias{transform_branch_length}
\title{Transform branch length for a re-scaled BM}
\usage{
transform_branch_length(phylo, alp)
}
\arguments{
\item{phylo:}{phylogenetic tree}

\item{alpha:}{a vector of values}
}
\value{
same phylogenetic tree, with transformed branch lengths

25/08/15 - Multivariate
}

