% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{update.allocate_regimes_from_shifts}
\alias{update.allocate_regimes_from_shifts}
\title{Update function for regime allocation.}
\usage{
update.allocate_regimes_from_shifts(edgeNbr, ancestral, shifts_edges, ...)
}
\arguments{
\item{edgeNbr}{: Number of the edge considered}

\item{ancestral}{: regime of the parent node}

\item{shifts_edges}{positions on edges}
}
\value{
regime of the daughter node.
}
\description{
\code{update.allocate_regimes_from_shifts} computes the regime of a daughter 
node, knowing the regime at the parent node and the vector of shifts positions
}
\details{
This function is used in function \code{allocate_regimes_from_shifts} and is 
designed to furnish function \code{recursionDown} with the right structure of
 data.
}
\keyword{internal}
