% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{update.compute_betas_from_shifts}
\alias{update.compute_betas_from_shifts}
\title{Update function for optimal value computation}
\usage{
update.compute_betas_from_shifts(edgeNbr, ancestral, shifts, ...)
}
\arguments{
\item{edgeNbr}{: Number of the edge considered}

\item{ancestral}{: Computed vector for the parental node}

\item{shifts}{position and values of the shifts}
}
\value{
Updated matrix of size (nNodes + ntaxa)x1.
}
\description{
\code{update.compute_betas_from_shifts} computes the optimal value at a daughter node, 
knowing the optimal value at the parent node and the vector of shifts.
}
\details{
This function is used in function \code{compute_betas_from_shifts} and is designed to 
furnish function \code{recursionDown} with the right structure of data.
}
\keyword{internal}
