% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{update.enumerate_parsimony}
\alias{update.enumerate_parsimony}
\title{Actualization of the enumeration.}
\usage{
update.enumerate_parsimony(daughters, daughtersParams, parent, cost, clus,
  pos, ...)
}
\arguments{
\item{daughters}{vector of daughters nodes.}

\item{daughtersParams}{list with length(daughters) entries, each entry being a 
list of k matrices representing the possible allocations starting from daughter.}

\item{parent}{the parent node.}
}
\value{
A list of size nclus, each entry being a matrix representing the possible 
allocations starting with node parent in state k.
}
\description{
\code{update.enumerate_parsimony} is used in function \code{enumerate_parsimony}, 
and compute the solution for the parent node, given its children.
}
\details{
This function takes a list with L entries corresponding to the children of a node,
and compute, for all the regimes, the possible allocations starting with parent
node in that regime. It uses functions \code{select.matrices} to select the
possible states of the children, and \code{matrix_of_possibles} to find the 
possible states.
}
\keyword{internal}
