% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{update.incidence.matrix}
\alias{update.incidence.matrix}
\title{Update function for incidence matrix}
\usage{
update.incidence.matrix(daughtersParams, ...)
}
\arguments{
\item{daughtersParams}{: rows of updated matrix corresponding to the
daughters of the current node.}
}
\value{
Vector of length ntaxa, indicating to which lineages the branch
above the current node belongs to.
}
\description{
\code{update.incidence.matrix} updates the matrix initialized in 
\code{init.incidence.matrix} for the computation of the incidence matrix
in \code{incidence.matrix}.
}
\details{
A node belongs to all the lineages of its daughters.
}
\keyword{internal}
