% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{update.parsimonyCost}
\alias{update.parsimonyCost}
\title{Actualization for parsimonyCost.}
\usage{
update.parsimonyCost(daughtersParams, ...)
}
\arguments{
\item{daughtersParams}{a (ndaughters) x (nclus) matrix with the line vectors of the cost
for the daughters node.}
}
\value{
A line vector corresponding to the parent node.
}
\description{
\code{update.parsimonyCost} compute the line vector of a parent node, given
the vectors of its daughters.
}
\details{
This function computes the cost of putting the parent in a state k, as the 
minimum number of shifts needed to get the given clustering of the trees bellow
parental node.
}
\keyword{internal}
