% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{update.parsimonyNumber}
\alias{update.parsimonyNumber}
\title{Actualization for parsimonyNumber.}
\usage{
update.parsimonyNumber(daughters, daughtersParams, cost, ...)
}
\arguments{
\item{daughters}{the identifiers of the daughters nodes.}

\item{daughtersParams}{a ndaughters x (nclus) matrix with the line vectors of the number of
solutions for the daughters nodes.}

\item{cost}{the (ntaxa + nNode) x nclus matrix of costs (computed by \code{parsimonyCost}).}
}
\value{
A line vector corresponding to the parent node.
}
\description{
\code{update.parsimonyNumber} compute the line vector of a parent node, given
the vectors of its daughters.
}
\details{
This function uses function \code{compute_state_filter} to find all the admissible
states of the daughters, given a starting state for the parent.
}
\keyword{internal}

