% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionsNumber.R
\name{update.partitionsNumber.gen}
\alias{update.partitionsNumber.gen}
\title{Update formula in the general case}
\usage{
update.partitionsNumber.gen(daughtersParams, ...)
}
\arguments{
\item{daughtersParams}{: matrix with 2*npart columns. Each row contains
the result of partitionsNumber for the children of a given node}
}
\value{
vector of size 2*npart. For k=1:npart it contains the number of
partitions with k groups compatible with the sub tree starting at the
current node ; and for k=(npart+1):2*npart, it contains the number of
"marqued" partitions with (k-npart) groups compatible with the sub tree
starting at the current node.
}
\description{
\code{update.partitionsNumber.gen} apply the actualisation formula to get
Nk and Ak of a node given its daughters.
}
\details{
Uses functions \code{sum.partitions} and \code{sum.simplex} to compute
the needed sums.
}
\keyword{internal}
