% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEM.R
\name{check_data}
\alias{check_data}
\title{Test the format of data entry.}
\usage{
check_data(phylo, Y_data, check.tips.names)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}

\item{Y_data}{matrix of data at the tips (pxntaxa)}

\item{check.tips.names}{(bool) whether to check the tips names or not}
}
\value{
Y_data a re-ordered matrix of data (if necessary)
}
\description{
\code{check_data} tests if the data matrix has the right format, and if it is
correctly ordered to match the tips names.
}
\keyword{internal}
