% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_step.R
\name{compute_tree_correlations_matrix.scOU}
\alias{compute_tree_correlations_matrix.scOU}
\title{Matrix of tree-induced correlations for the scOU}
\usage{
compute_tree_correlations_matrix.scOU(times_shared, distances_phylo, params_old,
  ...)
}
\arguments{
\item{times_shared}{times of shared ancestry of all nodes and tips, result of function
\code{compute_times_ca}}

\item{distances_phylo}{(matrix) : phylogenetics distance, result of function 
\code{compute_dist_phy}}

\item{params_old}{(list) : old parameters to be used in the E step}
}
\value{
matrix of variance covariance for the scOU
}
\description{
\code{compute_tree_correlations_matrix.scOU} computes the (n+m)x(m+n) matrix of correlations
induced by the tree. It takes two cases in consideration: root fixed, or root in stationary
state.
}
\keyword{internal}
