% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{correspondenceEdges}
\alias{correspondenceEdges}
\title{Correspondence between edges numbers}
\usage{
correspondenceEdges(edges, from, to)
}
\arguments{
\item{edges}{vector of index of edges in the tree "from"}

\item{from}{initial input tree (format "\code{phylo}")}

\item{to}{aimed output tree (format "\code{phylo}")}
}
\value{
vector of index of edges in the tree "to"
}
\description{
\code{correspondenceEdges} takes edges numbers on an input tree, and gives back
their corresponding numbers on the output tree.
}
