% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract.parsimonyNumber}
\alias{extract.parsimonyNumber}
\title{Extraction of the actual number of solutions.}
\usage{
\method{extract}{parsimonyNumber}(obj, node = attr(obj$nbrReconstructions,
  "ntaxa") + 1)
}
\arguments{
\item{obj}{an object of class "\code{parsimonyNumber}", result of function
\code{\link{parsimonyNumber}}.}

\item{node}{the root node of the subtree. By default, the root of the tree.}
}
\value{
An integer giving the number of equivalent parsimonious solutions.
}
\description{
\code{extract.parsimonyNumber} takes the two matrices computed by 
\code{\link{parsimonyNumber}}, and compute the actual number of parsimonious
solution for any subtree starting from a given node.
}
\details{
The parsimonious solutions are the one with the minimum number of shifts (that
are given by matrice costReconstructions). This function sums the number of 
solutions (given in matrice nbrReconstructions) that have the minimum number of 
shifts.
}

