% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{extract}
\alias{extract}
\title{Extraction function}
\usage{
extract(x, ...)
}
\arguments{
\item{x}{an S3 object.}

\item{...}{further arguments to be passed to the specific method.}
}
\value{
An integer giving the number of equivalent parsimonious solutions.
}
\description{
\code{extract} the needed quantities out of an S3 object.
}
\seealso{
\code{\link{extract.parsimonyNumber}},
\code{\link{extract.parsimonyCost}},
\code{\link{extract.enumerate_parsimony}},
\code{\link{extract.partitionsNumber}}
}
