% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{incidence.matrix.full}
\alias{incidence.matrix.full}
\title{Incidence matrix of a tree.}
\usage{
incidence.matrix.full(phylo)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}
}
\value{
Matrix of incidence, size ntaxa + Nnode.
}
\description{
\code{incidence.matrix.full} computes the incidence matrix U of a tree : for a 
node i and a branch b, U[i,b]=1 if b is in the lineage i, and 0 
otherwise.
}
\seealso{
\code{\link{incidence.matrix}}
}
