% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{incidence.matrix}
\alias{incidence.matrix}
\title{Incidence matrix of a tree.}
\usage{
incidence.matrix(phylo)
}
\arguments{
\item{phylo}{a phylogenetic tree, class \code{\link[ape]{phylo}}.}
}
\value{
Matrix of incidence, size nEdges x ntaxa.
}
\description{
\code{incidence.matrix} computes the incidence matrix T of a tree : for a 
lineage i and a branch b, T[i,b]=1 if b is in the lineage i, and 0 
otherwise.
}
\seealso{
\code{\link{incidence.matrix.full}}
}

