% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonyNumber.R
\name{init.enumerate_parsimony}
\alias{init.enumerate_parsimony}
\title{Initialization for the enumeration of parsimonious solutions.}
\usage{
init.enumerate_parsimony(phy, clusters, pos)
}
\arguments{
\item{phy}{Input tree.}

\item{clusters}{a vector representing the group of each tip.}
}
\value{
A list of size nNode + ntaxa, as described above.
}
\description{
\code{init.enumerate_parsimony} is used in function \code{enumerate_parsimony}, 
and initialize the correct data structure.
}
\details{
This function returns a list with nNodes + ntaxa entries. Entries correponding to
the tips are initialized with a list of nclus matrices. For tip i of group k, all
matrices are set to NULL, exept for the kth, set to a vector of size
nNodes + ntaxa, with entry i set to k, and all the others to NA.
}
\keyword{internal}

