% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutoff.R
\name{is.in.ranges}
\alias{is.in.ranges}
\title{Check whether parameters are in ranges.}
\usage{
is.in.ranges(p, min, max)
}
\arguments{
\item{p}{list of parameters with the correct structure}

\item{min}{list of minimum values for the parameters}

\item{max}{list of maximum values for the parameters}
}
\value{
boolean
}
\description{
\code{is.in.ranges.params} checks whether calculated parameters in the EM are
in the defined ranges.
}
\details{
This function is used to test the convergence of the algorithm.
}
\keyword{internal}
