% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{sample_edges_intervals}
\alias{sample_edges_intervals}
\title{Sample equally espaced edges.}
\usage{
sample_edges_intervals(tree, K)
}
\arguments{
\item{tree}{: imput tree}

\item{K}{: number of edges to be sampled.}
}
\value{
vector of edges
}
\description{
\code{sample_edges_intervals} samples K shifts, each in one of the K intervalles
regularly espaced on the height of the tree.
}
\details{
In case where the tree is not ultrametric, its "height" is defined as the minimum
tip height.
}
\keyword{internal}
