% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_manipulations.R
\name{shifts.matrix_to_list}
\alias{shifts.matrix_to_list}
\title{Compute the list of shifts.}
\usage{
shifts.matrix_to_list(delta)
}
\arguments{
\item{delta}{matrix description of the shift.}
}
\value{
List describing shifts.
}
\description{
\code{shifts.matrix_to_list} takes the vectorial description of the shifts 
to create the list description of the shifts.
}
\seealso{
\code{\link{shifts.list_to_matrix}}
}

