% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{transform_branch_length}
\alias{transform_branch_length}
\title{Transform branch length for a re-scaled BM}
\usage{
transform_branch_length(phylo, alp)
}
\arguments{
\item{phylo}{A phylogenetic tree of class \code{\link[ape]{phylo}}, with branch
lengths.}

\item{alp}{Value of the selection strength.}
}
\value{
phylo The same phylogenetic tree, with transformed branch lengths.
}
\description{
Re-scale the branch length of the tree so that a BM running
on the new tree produces the same observations at the tips than an OU with
parameter alpha.
}
